refactor(sigv4_helper.py): remove hooks.py module and move hooks to sigv4_helper.py
This refactor was needed in order to avoid a circular depdency, which resulted in a mid-module import.
